# api/openapi.yaml
openapi: "3.0.0"
info:
  title: GPIO
  description: The "GPIO" provides a REST based http interface providing the same functionality as the user
      interface. <br> To try it out within this interactive documentation, please customize the server variables below and
      authorize first. <br> <br> Please note that you have to switch the server to 'Basic App' before trying out the subscriber
      and the log api.
  contact:
      name: Support
      email: info@menucha.de
  version: 1.0.0
servers:
-  url: https://{mica-host}/{container-name}/rest/gpio
   description: GPIO App
   variables:
      mica-host:
         description: The ip of the MICA
         default: 10.10.10.10
      container-name:
         description: The name of the container
         default: GPIO
-  url: https://{mica-host}/{container-name}/rest
   description: Basic App
   variables:
      mica-host:
         description: The ip of the MICA
         default: 10.10.10.10
      container-name:
         description: The name of the container
         default: GPIO
security:
-  basic_auth: []
components:
  schemas:
    Device:
         type: object
         properties:
            id:
               type: string
            name:
               type: string
            label:
               type: string
            usable:
               type: boolean
            customized:
               type: boolean
            properties:
               type: object
               additionalProperties:
                  type: string
            fields:
               type: object
               additionalProperties:
                 $ref: '#/components/schemas/Field'
...
    Field:
         type: object
         properties:
            id:
               type: string
            name:
               type: string
            label:
               type: string
            properties:
               type: object
               additionalProperties:
                  type: string
            value:
               type: object
    Subscriptor:
         type: object
         properties:
            id:
               type: string
            enable:
               type: boolean
            name:
               type: string
            path:
               type: string
            subscriberId:
               type: string
            properties:
               type: object
               additionalProperties:
                  type: string
...
  securitySchemes:
    basic_auth:
       type: http
       scheme: basic             
paths:
...
  /devices/{id}:
      get:
         tags:
         - Device
         summary: Returns the gpio configuration with the given id as a JSON structure
         operationId: getDevice
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Device'
  /devices/{id}/label:
      get:
         tags:
         - Device
         summary: Returns the device label
         operationId: getLabel
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  text/plain:
                   schema:
                     type: string
      put:
         tags:
         - Device
         summary: Sets the device label
         operationId: setLabel
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         requestBody:
            content:
               text/plain:
                  schema:
                     type: string
         responses:
            '204':
               description: label updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
      
      delete:
        tags:
        - Device
        summary: Deletes the device label
        operationId:  removeDeviceLabel
        parameters: 
        - name: id
          in: path
          required: true
          schema:
            type: string
            
        responses:
            '204':
               description: Label deleted
            '500':
               description: Unexpected error occured
  /devices/{id}/properties:
      get:
         tags:
         - Properties
         summary: Returns the device properties
         operationId: getProperties
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
        
         responses:
            default:
               description: default response
               content:
                  application/json:
                    schema:
                      type: object
                      additionalProperties:
                        type: string
  /devices/{id}/properties/{name}:
      get:
         tags:
         - Properties
         summary: Returns the value of device property with the given name
         operationId: getProperty
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: name
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  text/plain:
                   schema:
                     type: string
                     
      put:
         tags:
         - Properties
         summary: Sets the value of device property with the given name
         operationId: setProperty
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: name
            in: path
            required: true
            schema:
               type: string         
         requestBody:
            content:
               text/plain:
                  schema:
                     type: string
         responses:
            '204':
               description: property updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
  /devices/{id}/fields:
      get:
         tags:
         - Fields
         summary: Returns a map with device fields
         operationId: getFields
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                    schema:
                      type: object
                      additionalProperties:
                        $ref: '#/components/schemas/Field'
  /devices/{id}/fields/{fieldId}:
      get:
         tags:
         - Fields
         summary: Returns the device field with the given fieldId
         operationId: getField
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                   schema:
                      $ref: '#/components/schemas/Field'
                     
      put:
         tags:
         - Fields
         summary: Sets the device field with the given fieldId
         operationId: setField
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string         
         requestBody:
            content:
              application/json:
                  schema:
                      $ref: '#/components/schemas/Field'
         responses:
            '204':
               description: field updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
  /devices/{id}/fields/{fieldId}/label:
      get:
         tags: 
         - Fields
         summary: Returns the field label
         operationId: getFieldLabel
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  text/plain:
                   schema:
                     type: string
      put:
         tags:
         - Fields
         summary: Set the field label
         operationId: setFieldLabel
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         requestBody:
            content:
               text/plain:
                  schema:
                     type: string
         responses:
            '204':
               description: label updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
      delete:
        tags:
        - Fields
        summary: Deletes the field label
        operationId:  removeFieldLabel
        parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
        responses:
            '204':
               description: Label deleted
            '500':
               description: Unexpected error occured
  /devices/{id}/fields/{fieldId}/value:
      get:
         tags: 
         - Fields
         summary: Returns the field value
         operationId: getFieldValue
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  text/plain:
                   schema:
                     type: string
      put:
         tags:
         - Fields
         summary: Set the field value
         operationId: setFieldValue
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         requestBody:
            content:
               text/plain:
                  schema:
                     type: string
         responses:
            '204':
               description: value updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
  /devices/{id}/fields/{fieldId}/properties:
      get:
         tags:
         - Fields
         summary: Returns the field properties
         operationId: getFieldProperties
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: fieldId
            in: path
            required: true
            schema:
              type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                    schema:
                      type: object
                      additionalProperties:
                        type: string
  /devices/{id}/fields/{fieldId}/properties/{name}:
      get:
         tags:
         - Fields
         summary: Returns the field property value
         operationId: getFieldProperty
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         -  name: name
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  text/plain:
                   schema:
                     type: string
                     
      put:
         tags:
         - Fields
         summary: Sets the field property
         operationId: setFieldProperty
         parameters:
         -  name: id
            in: path
            required: true
            schema:
              type: string
         -  name: fieldId
            in: path
            required: true
            schema:
               type: string
         -  name: name
            in: path
            required: true
            schema:
               type: string         
         requestBody:
            content:
               text/plain:
                  schema:
                     type: string
         responses:
            '204':
               description: property updated
            '415':
               description: Unsupported Media Type
            '422':
               description: Empty content not allowed
            '500':
               description: Unexpected error occured
...
  /reports/{id}/subscriptions:
      get:
         tags:
         - Reports
         summary: Returns a list of defined subscriptions for the given name
         operationId: getSubscriptors
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                     schema:
                        type: object
                        additionalProperties:
                           $ref: '#/components/schemas/Subscriptor'
      post:
         tags:
         - Reports
         summary: Adds a subscription for the given name
         operationId: defineSubscriptor
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Subscriptor'
         responses:
            default:
               description: default response
               content:
                  text/plain:
                     schema:
                        type: string
  /reports/{id}/subscriptions/{subscriptorId}:
      get:
         tags:
         - Reports
         summary: Returns the subscription with the given id of the given report as a JSON structure
         operationId: getSubscriptor
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: subscriptorId
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Subscriptor'
      put:
         tags:
         - Reports
         summary: Updates the subscription with the given id of the given report
         operationId: updateSubscriptor
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: subscriptorId
            in: path
            required: true
            schema:
               type: string
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Subscriptor'
         responses:
            default:
               description: default response
               content:
                  '*/*': {}
      delete:
         tags:
         - Reports
         summary: Deletes the subscription with the given id of the given report
         operationId: undefineSubscriptor
         parameters:
         -  name: id
            in: path
            required: true
            schema:
               type: string
         -  name: subscriptorId
            in: path
            required: true
            schema:
               type: string
         responses:
            default:
               description: default response
               content:
                  '*/*': {}
...